+++ wgpu-core/src/limits.rs
@@ -332,7 +332,8 @@ const UPLEVEL: Bucket = Bucket {
         subgroup_min_size: 4,
         subgroup_max_size: 128,
     },
-    features: Features::DEPTH_CLIP_CONTROL
+    features: Features::CORE_FEATURES_AND_LIMITS
+        .union(Features::DEPTH_CLIP_CONTROL)
         .union(Features::DEPTH32FLOAT_STENCIL8)
         // omit TEXTURE_COMPRESSION_ASTC
         // omit TEXTURE_COMPRESSION_ASTC_SLICED_3D
@@ -558,7 +559,8 @@ mod tests {
     #[test]
     fn enumerate_webgpu_features() {
         let difference = Features::all_webgpu_mask().difference(
-            Features::DEPTH_CLIP_CONTROL
+            Features::CORE_FEATURES_AND_LIMITS
+                .union(Features::DEPTH_CLIP_CONTROL)
                 .union(Features::DEPTH32FLOAT_STENCIL8)
                 .union(Features::TEXTURE_COMPRESSION_ASTC)
                 .union(Features::TEXTURE_COMPRESSION_ASTC_SLICED_3D)