+++ CHANGELOG.md
@@ -103,6 +103,8 @@ By @sagudev in [#10109](https://github.com/gfx-rs/wgpu/pull/10109).
 
   By @ErichDonGubler in [#8764](https://github.com/gfx-rs/wgpu/pull/8764)
 
+- Added `wgt::Features::CORE_FEATURES_AND_LIMITS` that are currently always reported as supported if device is WebGPU compliant as determined by [`wgpu_types::DownlevelCapabilities::is_webgpu_compliant`]. By @sagudev in [#10293](https://github.com/gfx-rs/wgpu/pull/10293).
+
 #### Naga
 
 - Add `@builtin(hit_barycentrics)`, a `vec2<f32>` readable in `@any_hit` and `@closest_hit` ray tracing pipeline shaders, holding two of the barycentric coordinates of the hit point on the triangle (the third is `1.0 - x - y`). Currently only supported with the SPIR-V backend. By @JMS55 in [#10193](https://github.com/gfx-rs/wgpu/pull/10193).